Move skill reference masking to SkillResource
